Setup for go based bpfd-agent
Rearrange directory structure so
that it's easy to have separate
binaries for the:
- bpfd-operator
- bpfd-agent
